Moved the duplicated export methods to a trait.
